+++ lib/src/options.dart
@@ -38,10 +38,12 @@ class PreprocessorOptions {
   final bool useColors;
 
   /** File to process by the compiler. */
-  String inputFile;
+  final String inputFile;
 
-  // We could make this faster, if it ever matters.
-  factory PreprocessorOptions() => parse(['']);
+  const PreprocessorOptions({this.verbose: false, this.checked: false,
+      this.lessSupport: true, this.warningsAsErrors: false,
+      this.throwOnErrors: false, this.throwOnWarnings: false,
+      this.useColors: true, this.polyfill: false, this.inputFile});
 
   PreprocessorOptions.fromArgs(ArgResults args)
       : warningsAsErrors = args['warnings_as_errors'],