+++ test/extend_test.dart
@@ -7,8 +7,6 @@ library extend_test;
 import 'package:unittest/unittest.dart';
 import 'testing.dart';
 
-var options = ['--warnings_as_errors', '--no-colors', 'memory'];
-
 compileAndValidate(String input, String generated) {
   var errors = [];
   var stylesheet = compileCss(input, errors: errors, opts: options);